requirements.txt
asgiref==3.8.1
channels==4.2.2
channels-redis==4.2.1
django==5.2.3
djangorestframework==3.16.0
msgpack==1.1.1
pip==25.1.1
pyjwt==2.9.0
python-dotenv==1.1.1
redis==6.2.0
ruff==0.12.2
sqlparse==0.5.3
